添加物料返回id

+++ buluerp_vue/ruoyi-admin/src/main/java/com/ruoyi/web/service/IErpMaterialInfoService.java
@@ -9,7 +9,7 @@ public interface IErpMaterialInfoService {
     List<ErpMaterialInfo> selectErpMaterialInfoList(ErpMaterialInfo erpMaterialInfo);
     List<ErpMaterialInfo> selectErpMaterialInfoListByIds(Long[] ids);
     ErpMaterialInfo selectErpMaterialInfoById(Long id);
-    int insertErpMaterialInfo(ErpMaterialInfo erpMaterialInfo) throws IOException;
+    Long insertErpMaterialInfo(ErpMaterialInfo erpMaterialInfo) throws IOException;
     int insertErpMaterialInfos(List<ErpMaterialInfo> erpMaterialInfos);
     int updateErpMaterialInfo(ErpMaterialInfo erpMaterialInfo) throws IOException;
     int deleteErpMaterialInfoByIds(Long[] ids);
